refactor slideshow
